+++ fastly/logging_https_test.go
@@ -53,7 +53,7 @@ Wm7DCfrPNGVwFWUQOmsPue9rZBgO
 			HeaderValue:       ToPointer("ExampleValue"),
 			Method:            ToPointer(http.MethodPut),
 			JSONFormat:        ToPointer("2"),
-			Period:            ToPointer(3600),
+			Period:            ToPointer(5),
 			Placement:         ToPointer("waf_debug"),
 			TLSCACert:         ToPointer(caCert),
 			TLSClientCert:     ToPointer(clientCert),
@@ -81,7 +81,7 @@ Wm7DCfrPNGVwFWUQOmsPue9rZBgO
 			HeaderValue:       ToPointer("ExampleValue"),
 			Method:            ToPointer(http.MethodPut),
 			JSONFormat:        ToPointer("2"),
-			Period:            ToPointer(3600),
+			Period:            ToPointer(5),
 			Placement:         ToPointer("waf_debug"),
 			TLSCACert:         ToPointer(caCert),
 			TLSClientCert:     ToPointer(clientCert),
@@ -110,7 +110,7 @@ Wm7DCfrPNGVwFWUQOmsPue9rZBgO
 			HeaderValue:       ToPointer("ExampleValue"),
 			Method:            ToPointer(http.MethodPut),
 			JSONFormat:        ToPointer("2"),
-			Period:            ToPointer(3600),
+			Period:            ToPointer(5),
 			Placement:         ToPointer("waf_debug"),
 			TLSCACert:         ToPointer(caCert),
 			TLSClientCert:     ToPointer(clientCert),
@@ -141,7 +141,7 @@ Wm7DCfrPNGVwFWUQOmsPue9rZBgO
 			HeaderValue:       ToPointer("ExampleValue"),
 			Method:            ToPointer(http.MethodPut),
 			JSONFormat:        ToPointer("2"),
-			Period:            ToPointer(3600),
+			Period:            ToPointer(5),
 			Placement:         ToPointer("waf_debug"),
 			TLSCACert:         ToPointer(caCert),
 			TLSClientCert:     ToPointer(clientCert),
@@ -215,7 +215,7 @@ Wm7DCfrPNGVwFWUQOmsPue9rZBgO
 	if *httpsCreateResp1.JSONFormat != "2" {
 		t.Errorf("bad json_format: %q", *httpsCreateResp1.JSONFormat)
 	}
-	if *httpsCreateResp1.Period != 3600 {
+	if *httpsCreateResp1.Period != 5 {
 		t.Errorf("bad period: %q", *httpsCreateResp1.Period)
 	}
 	if *httpsCreateResp1.Placement != "waf_debug" {